split server.ts in two
