+++ Dockerfile
@@ -24,6 +24,7 @@ ARG APP_DIR
 ENV SRVDIR="${APP_DIR}"
 LABEL io.cryostat.component=console-plugin
 COPY --from=backend_build /usr/src/app/node_modules/ "${APP_DIR}"/node_modules
+COPY --from=backend_build /usr/src/app/dist/main.js "${APP_DIR}"
 COPY --from=backend_build /usr/src/app/dist/server.js "${APP_DIR}"
 COPY --from=frontend_build /usr/src/app/dist "${APP_DIR}"/html
-ENTRYPOINT [ "/usr/bin/bash", "-c", "/usr/bin/node ${SRVDIR}/server.js" ]
+ENTRYPOINT [ "/usr/bin/bash", "-c", "/usr/bin/node ${SRVDIR}/main.js" ]
