+++ backend/.dockerignore
@@ -1,2 +1,3 @@
+!main.js
 !server.js
 !node_modules/
