+++ backend/package.json
@@ -2,12 +2,12 @@
   "name": "backend",
   "version": "4.2.0-dev",
   "description": "",
-  "main": "server.js",
+  "main": "main.js",
   "scripts": {
     "build": "tsc",
     "build:noCheck": "tsc --noCheck",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
+    "start": "node dist/main.js"
   },
   "author": "",
   "license": "MIT",
